Create accessible data tables examples
Jason here are the tables you can implement on the accessible version page of these charts/maps. I tried to keep it simple, and added a class so there are table borders and alternating row colors so data may be easier to process. I don't know if we should put all the data in one page, so maybe we'll have four links to four different pages...? Anyway these tables should do the trick for you. Any questions let me know. :)